Code refractor
