+++ src/pages/user/404.tsx
@@ -1,0 +1,9 @@
+import { Center } from '@chakra-ui/react';
+
+export default function SomeRoute() {
+  return (
+    <Center h="80vh">
+      <p>User Not found 404</p>
+    </Center>
+  );
+}
